Separate plumbing from example code
Change `bin/index.js` to just print a message
Move example code from there to `bin/example.js`
